requirements.txt
aiogram==3.10.0
mysql-connector-python==8.2.0
python-dotenv==1.0.1